Implement Minix inode write-back

+++ sys/vfs/vfs.h
@@ -45,7 +45,7 @@ typedef struct fs_node {
     uint32_t flags;       // Includes the node type.
     uint64_t inode;       // This is device-specific - provides a way for a filesystem to identify files.
     off_t    length;      // Size of the file, in bytes (64-bit).
-    uint32_t impl;        // An implementation-defined number.
+    uintptr_t impl;       // An implementation-defined number.
     
     // Timestamps (64-bit for Year 2038 compliance)
     int64_t atime;        // Last access time
